pyproject.toml
[project]
name = "watson-rfdiffusion-2023-gaia"
version = "0.1.0"
description = "Gaia knowledge package: Watson et al. 2023 — De novo design of protein structure and function with RFdiffusion (Nature)"
readme = "README.md"
authors = [
    { name = "kunchen", email = "chenkun0228@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "gaia-lang>=0.2.5",
]

[tool.hatch.build.targets.wheel]
packages = ["src/watson_rfdiffusion_2023"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.gaia]
type = "knowledge-package"
uuid = "67967e95-ae49-46a4-89f4-5ebe651d7483"
